+++ genomicsurveillance/__init__.py
@@ -5,7 +5,7 @@ GenomicSurveillance
 Models to estimate the incidence of virus variants.
 """
 
-from .models import IndependentMultiLineage
+from .models import IndependentMultiLineage, MultiLineage
 from .utils import (
     aggregate_tensor,
     alias_lineages,
@@ -27,4 +27,5 @@ __all__ = [
     "epiestim_R",
     "epiestim_discretise_serial_interval",
     "IndependentMultiLineage",
+    "MultiLineage",
 ]